fix: reorganize and format code
