+++ src/components/events/RestaurantFieldsForm.tsx
@@ -1,7 +1,6 @@
-// RestaurantFieldsForm.tsx
 import React from 'react';
 import TextField from '@mui/material/TextField';
-import { RestaurantOption } from './RestaurantOptionsForm';
+import { RestaurantOption } from '../../types/Event';
 
 interface Props {
   index: number;
